fix: use multiple test suites
