+++ package.json
@@ -39,7 +39,7 @@
   ],
   "license": "ISC",
   "dependencies": {
-    "@frmscoe/frms-coe-lib": "^2.2.0",
+    "@frmscoe/frms-coe-lib": "^2.2.2",
     "@frmscoe/frms-coe-startup-lib": "^2.1.0",
     "@log4js-node/logstash-http": "^1.1.0",
     "arangojs": "^8.4.0",
